会员模块字段命名规范

+++ rpc/doc/sql/cms_member_report.sql
@@ -6,8 +6,8 @@ create table cms_member_report
     report_member_name varchar(100) not null comment '举报人',
     create_time        datetime     not null,
     report_object      varchar(100) not null,
-    report_status      int(1)       not null comment '举报状态：0->未处理；1->已处理',
-    handle_status      int(1)       not null comment '处理结果：0->无效；1->有效；2->恶意',
+    report_status tinyint not null comment '举报状态：0->未处理；1->已处理',
+    handle_status tinyint not null comment '处理结果：0->无效；1->有效；2->恶意',
     note               varchar(200) null
 )
     comment '用户举报表' charset = utf8;
