+++ rpc/doc/sql/cms_topic_category.sql
@@ -5,7 +5,7 @@ create table cms_topic_category
     name          varchar(100) not null,
     icon          varchar(500) not null comment '分类图标',
     subject_count int          not null comment '专题数量',
-    show_status   int(2)       not null,
+    show_status tinyint not null,
     sort          int          not null
 )
     comment '话题分类表' charset = utf8;
