+++ rpc/doc/sql/pms_product_attribute.sql
@@ -10,9 +10,9 @@ create table pms_product_attribute
     sort                          int          not null comment '排序字段：最高的可以单独上传图片',
     filter_type                   int(1)       not null comment '分类筛选样式：1->普通；1->颜色',
     search_type                   int(1)       not null comment '检索类型；0->不需要进行检索；1->关键字检索；2->范围检索',
-    related_status                int(1)       not null comment '相同属性产品是否关联；0->不关联；1->关联',
-    hand_add_status               int(1)       not null comment '是否支持手动新增；0->不支持；1->支持',
-    type                          int(1)       not null comment '属性的类型；0->规格；1->参数'
+    related_status  tinyint not null comment '相同属性产品是否关联；0->不关联；1->关联',
+    hand_add_status tinyint not null comment '是否支持手动新增；0->不支持；1->支持',
+    type            tinyint not null comment '属性的类型；0->规格；1->参数'
 )
     comment '商品属性参数表';
