+++ rpc/doc/sql/sms_flash_promotion_session.sql
@@ -5,7 +5,7 @@ create table sms_flash_promotion_session
     name        varchar(200) not null comment '场次名称',
     start_time  time         not null comment '每日开始时间',
     end_time    time         not null comment '每日结束时间',
-    status      int(1)       not null comment '启用状态：0->不启用；1->启用',
+    status tinyint not null comment '启用状态：0->不启用；1->启用',
     create_time datetime     not null comment '创建时间'
 )
     comment '限时购场次表';
