+++ rpc/doc/sql/ums_integration_consume_setting.sql
@@ -2,10 +2,10 @@ create table ums_integration_consume_setting
 (
     id                    bigint auto_increment
         primary key,
-    deduction_per_amount  int    not null comment '每一元需要抵扣的积分数量',
-    max_percent_per_order int    not null comment '每笔订单最高抵用百分比',
-    use_unit              int    not null comment '每次使用积分最小单位100',
-    coupon_status         int(1) not null comment '是否可以和优惠券同用；0->不可以；1->可以'
+    deduction_per_amount  int     not null comment '每一元需要抵扣的积分数量',
+    max_percent_per_order int     not null comment '每笔订单最高抵用百分比',
+    use_unit              int     not null comment '每次使用积分最小单位100',
+    coupon_status         tinyint not null comment '是否可以和优惠券同用；0->不可以；1->可以'
 )
     comment '积分消费设置';
